pyproject.toml
[tool.poetry]
name = "codespaces-handson"
version = "0.1.0"
description = ""
authors = ["Atsushi Morimoto <74th.tech@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
mysql-connector-python = "^8.0.31"
mysqlclient = "^2.1.1"
types-mysqlclient = "^2.1.5.1"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pyright = "^1.1.286"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
